requirements.txt
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
pyyaml>=6.0
wandb>=0.16.0
